RD-2090 Migrate MapsActions to TypeScript
