+++ widgets/common/src/MapsActions.ts
@@ -1,0 +1,29 @@
+import { template } from 'lodash';
+
+class MapsActions {
+    constructor(private toolbox: Stage.Types.Toolbox) {}
+
+    public isAvailable() {
+        const { urlTemplate } = Stage.Common.Consts.leaflet;
+        const testUrl = template(urlTemplate, { interpolate: /{([\s\S]+?)}/g })({ x: 0, y: 0, z: 0, r: '' });
+
+        const noCache = new Date().getTime();
+        return this.toolbox
+            .getInternal()
+            .doGet(testUrl, { noCache })
+            .then(() => true)
+            .catch(() => false);
+    }
+}
+
+declare global {
+    namespace Stage.Common {
+        /* eslint-disable-next-line import/prefer-default-export */
+        export { MapsActions };
+    }
+}
+
+Stage.defineCommon({
+    name: 'MapsActions',
+    common: MapsActions
+});
